+++ app/models/quick_script/dev/post2_form_counter_script.rb
@@ -10,7 +10,7 @@ module QuickScript
           {
             :label   => "カウンタ",
             :key     => :my_counter,
-            :type    => :integer,
+            :type    => :numeric,
             :default => @my_counter,
           },
         ]
